requirements.txt
streamlit
opencv-python
easyocr
exifread
mtcnn
transformers
torch
Pillow
numpy
python-dotenv
groq
